requirements.txt
scikit-learn
streamlit
joblib
tldextract
numpy
requests==2.32.3
python-whois
ipaddress
dnspython
python-dotenv
virustotal-python==0.1.2
diskcache
pandas
scikit-learn
matplotlib
seaborn
python-whois==0.8.0
diskcache==5.6.3
aiohttp==3.10.5
python-dotenv==1.0.1

